+++ src/crypto/crypto.h
@@ -46,7 +46,6 @@
 #include "hex.h"
 #include "span.h"
 #include "hash.h"
-#include "device/device_declare.hpp"
 extern "C" {
   #include "crypto-ops.h"
 }
@@ -155,17 +154,6 @@ namespace crypto {
     friend bool check_ring_signature(const hash &, const key_image &,
       const public_key *const *, std::size_t, const signature *);
   };
-
-  secret_key generate_keys(public_key &pub, secret_key &sec, const secret_key& recovery_key, bool recover, hw::device &hwdev);
-  secret_key generate_keys(public_key &pub, secret_key &sec, hw::device &hwdev);
-  bool secret_key_to_public_key(const secret_key &sec, public_key &pub, hw::device &hwdev);
-  bool generate_key_derivation(const public_key &key1, const secret_key &key2, key_derivation &derivation, hw::device &hwdev);
-  void derivation_to_scalar(const key_derivation &derivation, size_t output_index, ec_scalar &res, hw::device &hwdev) ;
-  bool derive_public_key(const key_derivation &derivation, size_t output_index, const public_key &base, public_key &derived_key, hw::device &hwdev);
-  void derive_secret_key(const key_derivation &derivation, size_t output_index, const secret_key &base, secret_key &derived_key, hw::device &hwdev);
-  bool derive_subaddress_public_key(const public_key &out_key, const key_derivation &derivation, std::size_t output_index, public_key &derived_key, hw::device &hwdev);
-  void generate_key_image(const public_key &pub, const secret_key &sec, key_image &image, hw::device &hwdev);
-
 
   /* Generate N random bytes
    */
