+++ src/cryptonote_basic/account.h
@@ -33,7 +33,6 @@
 #include "cryptonote_basic.h"
 #include "crypto/crypto.h"
 #include "serialization/keyvalue_serialization.h"
-#include "device/device_declare.hpp"
 
 namespace cryptonote
 {
