+++ src/cryptonote_basic/cryptonote_basic.h
@@ -49,6 +49,7 @@
 #include "misc_language.h"
 #include "tx_extra.h"
 #include "ringct/rctTypes.h"
+#include "device/device.hpp"
 
 namespace cryptonote
 {
@@ -437,7 +438,7 @@ namespace cryptonote
     static inline keypair generate(hw::device &hwdev)
     {
       keypair k;
-      generate_keys(k.pub, k.sec, hwdev);
+      hwdev.generate_keys(k.pub, k.sec);
       return k;
     }
   };
