+++ src/wallet/api/wallet_manager.cpp
@@ -37,7 +37,6 @@
 #include "common/updates.h"
 #include "version.h"
 #include "net/http_client.h"
-#include "device/device.hpp"
 #include <boost/filesystem.hpp>
 #include <boost/regex.hpp>
